requirements.txt
openai>=1.51.0
streamlit>=1.44.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
pypdf>=5.1.0
